# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SIC API
  version: 0.1.0
  description: Implementation of accounting system based in hordak and following SIC
    (Andrisani) definitions.
  contact:
    name: gastong256
paths:
...
  /api/v1/companies/:
    get:
      operationId: list_companies
      summary: List companies
      parameters:
      - in: query
        name: all
        schema:
          type: boolean
        description: Return the full list in one response.
      - name: ordering
        required: false
        in: query
        description: Which field to use when ordering the results.
        schema:
          type: string
      - in: query
        name: page
        schema:
          type: integer
      - name: search
        required: false
        in: query
        description: A search term.
        schema:
          type: string
      - in: query
        name: summary
        schema:
          type: string
          enum:
          - selector
        description: Return a lightweight item shape for selectors.
      tags:
      - companies
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCompanyList'
          description: ''
    post:
      operationId: create_company
      summary: Create a company
      tags:
      - companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyCreate'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
        '400':
          description: Validation error
...
  /api/v1/companies/{company_id}/journal/{entry_id}/:
    get:
      operationId: get_journal_entry
      description: Returns full detail of a journal entry including all lines.
      summary: Retrieve a journal entry
      parameters:
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: path
        name: entry_id
        schema:
          type: integer
        required: true
      tags:
      - journal
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryDetail'
          description: ''
        '409':
          description: Company must be opened before journal is available
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company or entry not found
...
  /api/v1/companies/{id}/:
    get:
      operationId: get_company
      summary: Retrieve a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
    put:
      operationId: update_company
      summary: Update a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyWrite'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
        '400':
          description: Validation error
        '403':
          description: Permission denied
        '409':
          description: Read-only demo company
    patch:
      operationId: partial_update_company
      summary: Partially update a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCompanyWrite'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
        '400':
          description: Validation error
        '403':
          description: Permission denied
        '409':
          description: Read-only demo company
    delete:
      operationId: delete_company
      summary: Delete a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No content
        '403':
          description: Permission denied
        '409':
          description: Company has protected accounting records
  /api/v1/companies/{id}/opening-entry/:
    post:
      operationId: create_company_opening_entry
      summary: Create the opening entry for an existing company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyOpeningEntry'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryDetail'
          description: ''
        '400':
          description: Validation error
        '403':
          description: Permission denied
        '404':
          description: Company not found
        '409':
          description: Company already opened, already has entries, or is read-only
...
  /api/v1/items:
    post:
      operationId: create_item
      summary: Create item
      tags:
      - items
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemCreate'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
          description: ''
        '400':
          description: Validation error
  /api/v1/items/{id}:
    get:
      operationId: get_item
      summary: Get item by ID
      parameters:
      - in: path
        name: id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - items
      security:
      - jwtAuth: []
      - cookieAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
          description: ''
        '404':
          description: Item not found
...
components:
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        description:
          type: string
          description: Optional business description for a more realistic simulation.
        tax_id:
          type: string
          description: Simulated CUIT (optional).
          maxLength: 20
        owner_username:
          type: string
          readOnly: true
        account_count:
          type: integer
          readOnly: true
        books_closed_until:
          type: string
          format: date
          readOnly: true
          nullable: true
          description: Entries on or before this date are locked.
        is_demo:
          type: boolean
          readOnly: true
          description: Marks a seeded demo company shared across the system.
        is_read_only:
          type: boolean
          readOnly: true
          description: If true, accounting writes and metadata updates are blocked.
        has_opening_entry:
          type: boolean
          readOnly: true
        accounting_ready:
          type: boolean
          readOnly: true
        opening_entry_id:
          type: integer
          nullable: true
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - account_count
      - accounting_ready
      - books_closed_until
      - created_at
      - has_opening_entry
      - id
      - is_demo
      - is_read_only
      - name
      - opening_entry_id
      - owner_username
      - updated_at
    CompanyCreate:
      type: object
      properties:
        name:
          type: string
          description: Company name.
          maxLength: 255
        description:
          type: string
          default: ''
          description: Optional business description.
        tax_id:
          type: string
          default: ''
          description: Simulated CUIT (optional).
          maxLength: 20
        opening_entry:
          allOf:
          - $ref: '#/components/schemas/CompanyOpeningEntry'
          description: Optional balanced opening entry to create the company with
            initial capital/assets.
      required:
      - name
    CompanyOpeningAccountItem:
      type: object
      properties:
        name:
          type: string
          description: Movement-account name to create or reuse for this company.
          maxLength: 255
        parent_code:
          type: string
          description: Parent colectiva code in format X.XX (e.g. 1.01).
          maxLength: 4
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          description: Positive amount.
      required:
      - amount
      - name
      - parent_code
    CompanyOpeningEntry:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Accounting date of the opening entry.
        inventory_kind:
          allOf:
          - $ref: '#/components/schemas/InventoryKindEnum'
          default: INITIAL
          description: |-
            Whether the opening is based on the initial inventory or a later general inventory.

            * `INITIAL` - Inventario Inicial
            * `GENERAL` - Inventario General
        source_ref:
          type: string
          default: ''
          description: Optional reference, e.g. "APERTURA-001" or "INV-GRAL-001".
          maxLength: 100
        assets:
          type: array
          items:
            $ref: '#/components/schemas/CompanyOpeningAccountItem'
          description: Asset items that will be posted to Debe.
        liabilities:
          type: array
          items:
            $ref: '#/components/schemas/CompanyOpeningAccountItem'
          description: Liability items that will be posted to Haber.
      required:
      - assets
      - date
...
    CompanyWrite:
      type: object
      properties:
        name:
          type: string
          description: Company name.
          maxLength: 255
        description:
          type: string
          default: ''
          description: Optional business description.
        tax_id:
          type: string
          default: ''
          description: Simulated CUIT (optional).
          maxLength: 20
      required:
      - name
...
    InventoryKindEnum:
      enum:
      - INITIAL
      - GENERAL
      type: string
      description: |-
        * `INITIAL` - Inventario Inicial
        * `GENERAL` - Inventario General
    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          maxLength: 255
        description:
          type: string
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - created_at
      - id
      - name
    ItemCreate:
      type: object
      properties:
        name:
          type: string
          description: Human-readable item name.
          maxLength: 255
        description:
          type: string
          default: ''
          description: Optional description.
      required:
      - name
...
    JournalEntryDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        entry_number:
          type: integer
          readOnly: true
          title: Número de asiento
          description: Correlativo por empresa, asignado automáticamente.
        date:
          type: string
          format: date
          readOnly: true
          title: Fecha
        description:
          type: string
          readOnly: true
          title: Descripción
        source_type:
          allOf:
          - $ref: '#/components/schemas/SourceTypeEnum'
          readOnly: true
          title: Tipo de comprobante
        source_ref:
          type: string
          readOnly: true
          title: Referencia
          description: 'Ej.: "Factura A-0001", "Recibo 00123".'
        created_by:
          type: string
          readOnly: true
        reversal_of_id:
          type: integer
          readOnly: true
        reversed_by_id:
          type: integer
          nullable: true
          readOnly: true
        total_debit:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
        total_credit:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntryLineRead'
          readOnly: true
      required:
      - created_by
      - date
      - description
      - entry_number
      - id
      - lines
      - reversal_of_id
      - reversed_by_id
      - source_ref
      - source_type
      - total_credit
      - total_debit
    JournalEntryLineRead:
      type: object
      properties:
        account_id:
          type: integer
          title: Cuenta
          readOnly: true
        account_code:
          type: string
          readOnly: true
        account_name:
          type: string
          readOnly: true
        type:
          allOf:
          - $ref: '#/components/schemas/TypeEnum'
          readOnly: true
          title: Tipo
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
          title: Importe
      required:
      - account_code
      - account_id
      - account_name
      - amount
      - type
...
    PaginatedCompanyList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Company'
...
    PatchedCompanyWrite:
      type: object
      properties:
        name:
          type: string
          description: Company name.
          maxLength: 255
        description:
          type: string
          default: ''
          description: Optional business description.
        tax_id:
          type: string
          default: ''
          description: Simulated CUIT (optional).
          maxLength: 20
...
    SourceTypeEnum:
      enum:
      - OPENING
      - MANUAL
      - INVOICE
      - RECEIPT
      - OTHER
      type: string
      description: |-
        * `OPENING` - Apertura
        * `MANUAL` - Manual
        * `INVOICE` - Factura
        * `RECEIPT` - Recibo
        * `OTHER` - Otro
...
    TypeEnum:
      enum:
      - DEBIT
      - CREDIT
      type: string
      description: |-
        * `DEBIT` - Deudora
        * `CREDIT` - Acreedora
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT